projects/LAB-04/chavatte_image_processing/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
tifffile==2025.2.18